Use 'sources' to obtain SHA512 checksum of image
The lookaside cache does not support uploading with the SHA256 hash
algorithm, so instead we have opted to store a sources file in the
installer container at the path /usr/share/coreos-sources. Read that
file to get the SHA512 hash of the ISO, for use in the download URL for
the lookaside cache.