Don't calculate ISO checksum when extracting
Recalculating the checksum to see if the file is up to date makes
restarting the pod slow, and requires an additional copy if the ISO has
been modified to add kernel command line arguments. Instead, copy over
the SHA256 sum file and compare its contents to the latest SHA256 sum.