Acquire aarch64 ISO at runtime
Instead of baking the ISO into the image at build time, we extract it
when we run the copy-metal script.  This avoids increasing the release
payload unnecessarily.

We use the oc command to do the extraction and so we add to the image.
This patch is accompanied by changes to cluster-baremetal-operator to
provide the name of the machine-os-images as an env var and to mount the
the pull secret and the certificates.

Notably absent from this change is any support for downloading aarch64
images during bootstrapping.

(cherry picked from commit 73f5056705164b60578de8f64bd3b4398383a040)